Added more parameters. Smoothing for volume change. Bigger font.

+++ app/src/main/java/com/phantompowerracing/ict/AudioPlayer.java
@@ -111,8 +111,8 @@ public class AudioPlayer implements Runnable {
         //audioTrack.pause();
     }
 
-    public void setVolume(float vol) {
-        audioTrack.setVolume(vol);
+    public void setVolume(double vol) {
+        audioTrack.setVolume((float)vol);
     }
 
     HandlerThread callbackThread;
